Refactored controller and remove 'likes' actions.

+++ controllers/postController.js
@@ -1,40 +1,4 @@
-const { userModel, themeModel, postModel } = require('../models');
-
-/**
- * 
- * @param {String} text 
- * @param {String} userId 
- * @param {String} themeId 
- * @returns {void}
- */
-
-function newPost(text, userId, themeId) {
-    return postModel.create({ text, userId, themeId })
-        .then(post => {
-            return Promise.all([
-                userModel.updateOne({ _id: userId }, { $push: { posts: post._id }, $addToSet: { themes: themeId } }),
-                themeModel
-                    .findByIdAndUpdate(
-                        { _id: themeId }, 
-                        { $push: { posts: post._id }, 
-                        $addToSet: { subscribers: userId } }, 
-                        { new: true }
-                    )
-                    .populate({
-                        path: 'posts',
-                        select: {
-                            __v: 0
-                        },
-                        populate: {
-                          path: 'userId',
-                          select: {
-                            username: 1,
-                          }
-                        }
-                    })
-            ])
-        })
-}
+const { postModel } = require('../models');
 
 /**
  * 
@@ -47,25 +11,12 @@ function newPost(text, userId, themeId) {
 function getLatestsPosts(req, res, next) {
     const limit = Number(req.query.limit) || 0;
 
-    postModel
-        .find(
-            {}, 
-            { 
-                __v: 0, 
-                likes: 0 
-            }
-        )
+    postModel.find({})
         .sort({ createdAt: -1 })
         .limit(limit)
         .populate([
             {
-                path: 'themeId',
-                select: {
-                    themeName: 1
-                }
-            },
-            {
-                path: 'userId',
+                path: 'authorId',
                 select: {
                     username: 1
                 }
@@ -86,12 +37,11 @@ function getLatestsPosts(req, res, next) {
  */
 
 function createPost(req, res, next) {
-    const { themeId } = req.params;
     const { _id: userId } = req.user;
-    const { postText } = req.body;
+    const { text, themeId } = req.body;
 
-    newPost(postText, userId, themeId)
-        .then(([_, updatedTheme]) => res.status(200).json(updatedTheme))
+    postModel.create({ text, authorId: userId, themeId })
+        .then(x => { res.status(201).json(x) })
         .catch(next);
 }
 
@@ -105,14 +55,14 @@ function createPost(req, res, next) {
 
 function editPost(req, res, next) {
     const { postId } = req.params;
-    const { postText } = req.body;
+    const { text } = req.body;
     const { _id: userId } = req.user;
 
     // if the userId is not the same as this one of the post, the post will not be updated
-    postModel.findOneAndUpdate({ _id: postId, userId }, { text: postText }, { new: true })
+    postModel.findOneAndUpdate({ _id: postId, authorId: userId }, { text }, { new: true })
         .then(updatedPost => {
             if (updatedPost) {
-                res.status(200).json(updatedPost);
+                res.status(201).json(updatedPost);
             }
             else {
                 res.status(401).json({ message: `Not allowed!` });
@@ -130,17 +80,14 @@ function editPost(req, res, next) {
  */
 
 function deletePost(req, res, next) {
-    const { postId, themeId } = req.params;
+    const { postId } = req.params;
     const { _id: userId } = req.user;
 
-    Promise.all([
-        postModel.findOneAndDelete({ _id: postId, userId }),
-        userModel.findOneAndUpdate({ _id: userId }, { $pull: { posts: postId } }),
-        themeModel.findOneAndUpdate({ _id: themeId }, { $pull: { posts: postId } }),
-    ])
-        .then(([deletedOne, _, __]) => {
-            if (deletedOne) {
-                res.status(200).json(deletedOne)
+    // if the userId is not the same as this one of the post, the post will not be deleted
+    postModel.findOneAndDelete({ _id: postId, authorId: userId })
+        .then(deletedPost => {
+            if (deletedPost) {
+                res.status(200).json(deletedPost)
             } else {
                 res.status(401).json({ message: `Not allowed!` });
             }
@@ -148,50 +95,9 @@ function deletePost(req, res, next) {
         .catch(next);
 }
 
-/**
- * 
- * @param {Request} req 
- * @param {Response} res 
- * @param {Callback} next 
- * @returns {void}
- */
-
-function like(req, res, next) {
-    const { postId } = req.params;
-    const { _id: userId } = req.user;
-
-    // console.log('like');
-
-    postModel.updateOne({ _id: postId }, { $addToSet: { likes: userId } }, { new: true })
-        .then(() => res.status(200).json({ message: 'Liked successful!' }))
-        .catch(next);
-}
-
-/**
- * 
- * @param {Request} req 
- * @param {Response} res 
- * @param {Callback} next 
- * @returns {void}
- */
-
-function unlike(req, res, next) {
-    const { postId } = req.params;
-    const { _id: userId } = req.user;
-
-    // console.log('unlike');
-
-    postModel.updateOne({ _id: postId }, { $pull: { likes: userId } }, { new: true })
-        .then(() => res.status(200).json({ message: 'Unliked successfull!' }))
-        .catch(next);
-}
-
 module.exports = {
     getLatestsPosts,
-    newPost,
     createPost,
     editPost,
-    deletePost,
-    like,
-    unlike
+    deletePost
 }
